fix tests
The tests were broken on report-txt and report-svg.

The -txt report was comparing ansi-ified output to non-ansi-ified output,
where the ansi-ified output was coming from nscm.  Do I have chalk set wrong
or something locally?  Who knows.  I ended up stripping the ansi sequences
out of the nscm output, and now it passes the test.

For -svg, the output was _slightly_ different - like path coordinates were
different in the 0.1% diff kinda range.  F... Maybe this is a differently compiled
version of viz.js or something?  Visually, there was no difference between the
actual and expected graphs when displayed.  So I just updated the expected one
in fixtures to the newly output'd one.